adding wrapper (connect, select) for using mainValue as prop in DynamicZone/Component
