+++ packages/core/admin/admin/src/content-manager/components/DynamicZone/components/Component/utils/connect.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function connect(WrappedComponent, select) {
+  return function(props) {
+    const selectors = select(props);
+
+    return <WrappedComponent {...props} {...selectors} />;
+  };
+}
+
+export default connect;
