+++ packages/core/admin/admin/src/content-manager/components/DynamicZone/components/Component/utils/index.js
@@ -1,0 +1,2 @@
+export { default as connect } from './connect';
+export { default as select } from './select';
