+++ packages/core/admin/admin/src/content-manager/components/DynamicZone/components/Component/utils/select.js
@@ -1,0 +1,19 @@
+import { useMemo } from 'react';
+import useMainValue from '../hooks/useMainValue';
+import { useContentTypeLayout } from '../../../../../hooks';
+
+function useSelect({ componentUid, name, index }) {
+  const { getComponentLayout } = useContentTypeLayout();
+  const componentLayoutData = useMemo(() => {
+    const layout = getComponentLayout(componentUid);
+
+    return layout;
+  }, [componentUid, getComponentLayout]);
+  const mainValue = useMainValue(componentLayoutData, [name, index]);
+
+  return {
+    mainValue,
+  };
+}
+
+export default useSelect;
